Creating Resolution1 style file
